requirements.txt
boto3==1.26.95
cryptography==39.0.2
django-storages==1.13.2
gunicorn==20.1.0
mysqlclient==2.1.1
pip-chill==1.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
tzdata==2022.7
django-widget-tweaks==1.4.12
json-log-formatter==0.5.2
requests==2.28.2